migrations changes

+++ database/migrations/2023_01_13_151827_create_courses_comments_table.php
@@ -13,32 +13,34 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('courses_comments', function (Blueprint $table) {
-            $table->id()->primary();
+        if (!Schema::hasTable('courses_comments')) {
+            Schema::create('courses_comments', function (Blueprint $table) {
+                $table->id()->primary();
 
-            $table->bigInteger('id_course')->unsigned();
-            $table->foreign('id_course')
-                ->references('id')
-                ->on('courses')
-                ->constrained()
-                ->onDelete('cascade');
+                $table->bigInteger('id_course')->unsigned();
+                $table->foreign('id_course')
+                    ->references('id')
+                    ->on('courses')
+                    ->constrained()
+                    ->onDelete('cascade');
 
-            $table->bigInteger('id_user')->unsigned();
-            $table->foreign('id_user')
-                ->references('id')
-                ->on('users')
-                ->constrained()
-                ->onDelete('cascade');
+                $table->bigInteger('id_user')->unsigned();
+                $table->foreign('id_user')
+                    ->references('id')
+                    ->on('users')
+                    ->constrained()
+                    ->onDelete('cascade');
 
-            $table->bigInteger('id_comment')->unsigned();
-            $table->foreign('id_comment')
-                ->references('id')
-                ->on('comments')
-                ->constrained()
-                ->onDelete('cascade');
+                $table->bigInteger('id_comment')->unsigned();
+                $table->foreign('id_comment')
+                    ->references('id')
+                    ->on('comments')
+                    ->constrained()
+                    ->onDelete('cascade');
 
-            $table->timestamps();
-        });
+                $table->timestamps();
+            });
+        }
     }
 
     /**
