+++ database/migrations/2023_01_25_110148_create_followed_courses_table.php
@@ -13,23 +13,25 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('followed_courses', function (Blueprint $table) {
-            $table->id()->primary();
+        if (!Schema::hasTable('followed_courses')) {
+            Schema::create('followed_courses', function (Blueprint $table) {
+                $table->id()->primary();
 
-            $table->bigInteger('id_course')->unsigned();
-            $table->foreign('id_course')
-                ->references('id')
-                ->on('courses')
-                ->constrained()
-                ->onDelete('cascade');
+                $table->bigInteger('id_course')->unsigned();
+                $table->foreign('id_course')
+                    ->references('id')
+                    ->on('courses')
+                    ->constrained()
+                    ->onDelete('cascade');
 
-            $table->bigInteger('id_user')->unsigned();
-            $table->foreign('id_user')
-                ->references('id')
-                ->on('users')
-                ->constrained()
-                ->onDelete('cascade');
-        });
+                $table->bigInteger('id_user')->unsigned();
+                $table->foreign('id_user')
+                    ->references('id')
+                    ->on('users')
+                    ->constrained()
+                    ->onDelete('cascade');
+            });
+        }
     }
 
     /**
